1. Changed Knight to a functional component so that I could call an event listener for key input, however now it doesn't run in the gameloop so it's quite clunky. 2. Downloaded Xcode and tried to convert this into native code for android/ios, I d on't think it worked. Finally I would just like to say I think this solution for moving is quite poor, maybe making it back into an entity and having it rerender each time the loop runs would wo rk. I just kept getting an error Invalid hook call. Hooks can only be called inside of the body of a function component. This could happen for one of the following reasons:
You might have mismatching versions of React and the renderer (such as React DOM)
You might be breaking the Rules of Hooks
You might have more than one copy of React in the same app See for tips about how to debug and fix this problem. I believe it was number 2 and I wasn't sure how to work around it.